% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ari_palette}
\alias{ari_palette}
\title{An Ari Aster palette generator}
\usage{
ari_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Heriditary1}, \code{Heriditary2}, \code{Heridtary3}, \code{Heridtary4},
\code{Midsommar1}, \code{Midsommar2}, \code{Midsommar3}, \code{Midsommar4},
\code{Beau1}, \code{Beau2}, \code{Beau3}, \code{BeauMonster},
\code{Eddington1}, \code{Eddington2}, \code{Eddington3},}

\item{n}{Number of colors desired. Most palettes have 5 or 6 colors.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes inspired by Ari Aster films.
}
\examples{
ari_palette("Hereditary1")
ari_palette("Midsommar2")
ari_palette("Beau3")
ari_palette("Midsommar1", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- ari_palette(21, name = "Hereditary2", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
