% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ari_palettes}
\alias{ari_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
ari_palettes
}
\description{
Use \code{\link{ari_palette}} to construct palettes of desired length.
}
\keyword{datasets}
